[Refectoring]FadeingModel.tsx ImageFadeMaterial model 타입 오류 해결
