+++ src/types/declare.ts
@@ -1,0 +1,16 @@
+import { ImageFadeMaterial } from "@/utils/materials/ImageMaterial";
+import { RefObject } from "react";
+import * as THREE from "three";
+
+declare module "@react-three/fiber" {
+  interface ThreeElements {
+    imageFadeMaterial: THREE.ShaderMaterialParameters & {
+      origin?: THREE.Texture;
+      replace?: THREE.Texture;
+      effect?: THREE.Texture;
+      dispFactor?: number;
+      effectFactor?: number;
+      ref: RefObject<typeof ImageFadeMaterial>;
+    };
+  }
+}
